+++ awslimitchecker/connectable.py
@@ -50,6 +50,9 @@ class ConnectableCredentials(object):
     `boto3.sts.STSConnection.assume_role <https://boto3.readthedocs.org/en/
     latest/reference/services/sts.html#STS.Client.assume_role>`_ just returns
     a dict. This class provides a compatible interface for boto3.
+
+    We also maintain an ``account_id`` attribute that can be set to the
+    account ID, to ensure that credentials are updated when switching accounts.
     """
 
     def __init__(self, creds_dict):
@@ -59,7 +62,7 @@ class ConnectableCredentials(object):
         self.expiration = creds_dict['Credentials']['Expiration']
         self.assumed_role_id = creds_dict['AssumedRoleUser']['AssumedRoleId']
         self.assumed_role_arn = creds_dict['AssumedRoleUser']['Arn']
-        self.account_id = creds_dict['account_id']
+        self.account_id = None
 
 
 class Connectable(object):
@@ -98,11 +101,10 @@ class Connectable(object):
                     logger.debug("Reusing previous STS credentials for "
                                  "account %s", self.account_id)
                 else:
-                    logger.debug("Previous STS credentials are for account %s",
-                                 Connectable.credentials.account_id)
-                    logger.debug("Connecting for account %s role '%s' with STS "
-                                 "(region: %s)", self.account_id,
-                                 self.account_role, self.region)
+                    logger.debug("Previous STS credentials are for account %s; "
+                                 "getting new credentials for current account "
+                                 "(%s)", Connectable.credentials.account_id,
+                                 self.account_id)
                     Connectable.credentials = self._get_sts_token()
             kwargs['aws_access_key_id'] = Connectable.credentials.access_key
             kwargs['aws_secret_access_key'] = Connectable.credentials.secret_key
@@ -177,10 +179,9 @@ class Connectable(object):
             assume_kwargs['TokenCode'] = self.mfa_token
         role = sts.assume_role(**assume_kwargs)
 
-        role['account_id'] = self.account_id
-
         creds = ConnectableCredentials(role)
+        creds.account_id = self.account_id
 
-        logger.debug("Got STS credentials for role; access_key_id=%s",
-                     creds.access_key)
+        logger.debug("Got STS credentials for role; access_key_id=%s "
+                     "(account_id=%s)", creds.access_key, creds.account_id)
         return creds